# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KISY Enterprise Messenger API
  version: 0.4.0
  description: >
    REST API for the KISY enterprise messenger. Covers stages 2–4:
    authentication, access control, invitations, profile, groups, private
    chats, messages, reactions, read state, favorites, notifications, the
    CEO admin surface and the WebSocket gateway. File handling is added in a
    later stage.
servers:
  - url: /api/v1
    description: Versioned API root (behind the Nginx edge proxy)
...
paths:
...
  /messages/{messageID}/reactions:
    post:
      tags: [messages]
      summary: Add an emoji reaction (idempotent)
      parameters:
        - { name: messageID, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji: { type: string, maxLength: 32 }
      responses:
        '200': { description: Reaction added, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [messages]
      summary: Remove your emoji reaction
      parameters:
        - { name: messageID, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji: { type: string, maxLength: 32 }
      responses:
        '200': { description: Reaction removed, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /read:
    post:
      tags: [messages]
      summary: Mark a chat read up to a message (updates unread counters)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatType, chatId, messageId]
              properties:
                chatType: { type: string, enum: [private, group] }
                chatId: { type: string, format: uuid }
                messageId: { type: string, format: uuid }
      responses:
        '200': { description: Read position stored, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /notifications:
    get:
      tags: [notifications]
      summary: List notifications with unread count
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items: { $ref: '#/components/schemas/Notification' }
                          unreadCount: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/read:
    post:
      tags: [notifications]
      summary: Mark one notification (by id) or all read
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: [string, 'null'], format: uuid }
      responses:
        '200': { description: Marked read, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: kisy_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: kisy_refresh
...
  schemas:
    Envelope:
      type: object
      required: [success, requestId, timestamp]
      properties:
        success: { type: boolean }
        data: { type: object }
        error: { $ref: '#/components/schemas/Error' }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Error:
      type: object
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_TOKEN
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EXPIRED
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - RATE_LIMITED
            - VALIDATION_FAILED
            - INTERNAL_ERROR
        message: { type: string }
        details: {}
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        payload: { type: object }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
    NotFound:
      description: Resource not found (or hidden from this clearance)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
security:
  - accessCookie: []